Releasing version 0.1.0
This version includes base functionality to set lights

Lights can be set to:
* On
* Off
* Dimmed
* A custom RGB and Brightness value

Default responses are included, no meaningful information is returned to
a succesful request.

For deployment, the spring profile should be set to prod, and a custom
application-prod.properties file be given, containing the API key and
Hue bridge IP. The server port and logging level should also be defined,
but those are optional.